Refactor: Button 컴포넌트 CircleButton으로 변경
